api: rename CohortAdapter to CohortPort + clean kdocs
Three refinements with no behavioural change:

- Rename the interface from `CohortAdapter` to `CohortPort` and move
  it into a dedicated `cohort/port/` package. The implementation
  stays under `cohort/adapter/<vendor>/`, keeping the hexagonal
  port/adapter split visible at the package level. The previous
  `*Adapter` suffix on the interface was inconsistent with the
  outbound-port role and required a workaround to dodge the
  "production *Adapter must reside in ..adapter.." arch rule.
- Rename the job's lazy-resolve helper from `createExternal` to
  `findOrCreateExternalCohortId`. The new name mirrors the existing
  `findOrCreateList` idiom in this codebase and surfaces the verb
  ("createCohort") that the port itself uses, making the
  responsibilities of caller and port line up by vocabulary.
- Strip rollout-stage commentary from kdocs (PR-stage labels,
  "follow-up cutover", "two-phase rollout"). The kdocs now describe
  the contract and intent as if the code is the steady state, not a
  point in a migration timeline.

Legacy `BrevoListAdapter` and `ContactListAdapter` stay named as-is
in this PR: they are deleted outright in the engine-cutover PR and
renaming them in flight would be churn for code that does not
survive.
